Move embedding models
